Added Weightloss plot
Added the weightloss plot as a stand alone thing. It is inside a div so
it can be moved where ever it needs to go. It can also be resized in
terms of the div. Also, right now it is set up to query a REST server
however, none of the specifics beyond that are done yet.